+++ platforms/reference/src/ReferencePyTorchKernelsE2EDiffConf.cpp
@@ -100,7 +100,7 @@ void ReferenceCalcPyTorchForceE2EDiffConfKernel::initialize(const System& system
 	vector<vector<int>> tmpPairs = force.getPairs();
 	vector<vector<int>> tmpTetras = force.getTetras();
 	vector<vector<int>> tmpCisTrans = force.getCisTrans();
-	vector<vector<float>> tmpEncoding = force.getEncoding();
+	vector<vector<double>> tmpEncoding = force.getEncoding();
 	usePeriodic = force.usesPeriodicBoundaryConditions();
 
 		
@@ -204,7 +204,7 @@ void ReferenceCalcPyTorchForceE2EDiffConfKernel::initialize(const System& system
 	// encoding
 	for (int i = 0; i < tmpEncoding.size(); i++) {
 		for (int j = 0; j < tmpEncoding[i].size(); j++) {
-			enc_acc[i][j] = tmpEncoding[i][j];
+		  enc_acc[i][j] = float(tmpEncoding[i][j]);
 		}
 	}
